Tweaks to make it work with most servos.
In particular, it targets a specific pulse timing for the full
range of the input signal. The useful range for most servos is
about 1ms to 2ms of a pulse. The code calculates the desired
pulse length and then converts this into the PWM timing
necessary. This uses floating point for now, we may want to use
integer math later for performance.

We drop the processor clock to 1MHz and set the PWM counter such that
we get something like a 16ms PWM period. This may be too quick for
some servos but seems to work for this.

We also setup some registers that were missing.